update all deps use new source set structure and cleanup some of the warnings
