+++ build.gradle.kts
@@ -1,8 +1,7 @@
 import org.apache.tools.ant.taskdefs.condition.Os
 
 plugins {
-    id("dev.petuska.npm.publish") version "3.4.1"
-    kotlin("multiplatform") version "1.9.23"
+    kotlin("multiplatform") version "1.9.24"
     id("com.android.library")
     id("io.codearte.nexus-staging") version "0.30.0"
     `maven-publish`
@@ -39,140 +38,63 @@ kotlin {
     androidTarget {
         publishLibraryVariants("release")
     }
-    if (loadAllPlatforms) {
-        jvm {
-            compilations.all {
-                kotlinOptions.jvmTarget = "1.8"
-            }
-            testRuns["test"].executionTask.configure {
-                useJUnit()
-            }
+    jvm {
+        compilations.all {
+            kotlinOptions.jvmTarget = "1.8"
         }
-        js(IR) {
-            moduleName = "buffer-kt"
-            browser { binaries.library() }
-            nodejs {
-                binaries.library()
-                testTask {
-                    useMocha {
-                        timeout = "10s"
-                    }
+        testRuns["test"].executionTask.configure {
+            useJUnit()
+        }
+    }
+    js {
+        moduleName = "buffer-kt"
+        browser { binaries.library() }
+        nodejs {
+            binaries.library()
+            testTask {
+                useMocha {
+                    timeout = "10s"
                 }
             }
         }
     }
-
     macosX64()
     macosArm64()
     linuxX64()
     linuxArm64()
-    ios()
+    iosArm64()
     iosSimulatorArm64()
-    watchos()
+    watchosArm64()
     watchosSimulatorArm64()
-    tvos()
+    tvosArm64()
     tvosSimulatorArm64()
-
+    applyDefaultHierarchyTemplate()
     sourceSets {
-        val commonMain by getting
-        val commonTest by getting {
-            dependencies {
-                implementation(kotlin("test"))
-                implementation("org.jetbrains.kotlinx:kotlinx-coroutines-core:1.8.1")
-            }
+        commonTest.dependencies {
+            implementation(kotlin("test"))
+            implementation("org.jetbrains.kotlinx:kotlinx-coroutines-core:1.8.1")
         }
-        val androidMain by getting {
-            dependencies {
-                implementation("org.jetbrains.kotlinx:kotlinx-coroutines-core:1.8.1")
-            }
+        androidMain.dependencies {
+            implementation("org.jetbrains.kotlinx:kotlinx-coroutines-core:1.8.1")
         }
+
         val androidUnitTest by getting {
             kotlin.srcDir("src/commonJvmTest/kotlin")
         }
-        val androidInstrumentedTest by getting {
-            dependsOn(commonTest)
-            kotlin.srcDir("src/commonJvmTest/kotlin")
-            kotlin.srcDir("src/commonTest/kotlin")
-            dependencies {
-                implementation("androidx.test:runner:1.5.2")
-                implementation("androidx.test:rules:1.5.0")
-                implementation("androidx.test:core-ktx:1.5.0")
-            }
-        }
-        if (loadAllPlatforms) {
-            val jvmMain by getting
-            val jvmTest by getting {
-                kotlin.srcDir("src/commonJvmTest/kotlin")
-            }
+//        val androidInstrumentedTest by getting {
+//            dependsOn(commonTest)
+//            kotlin.srcDir("src/commonJvmTest/kotlin")
+//            kotlin.srcDir("src/commonTest/kotlin")
+//            dependencies {
+//                implementation("androidx.test:runner:1.5.2")
+//                implementation("androidx.test:rules:1.5.0")
+//                implementation("androidx.test:core-ktx:1.5.0")
+//            }
+//        }
 
-            val jsMain by getting {
-                dependencies {
-                    implementation("org.jetbrains.kotlin-wrappers:kotlin-web:1.0.0-pre.746")
-                    implementation("org.jetbrains.kotlin-wrappers:kotlin-js:1.0.0-pre.746")
-                }
-            }
-            val jsTest by getting
-        }
-        val macosX64Main by getting
-        val macosX64Test by getting
-        val macosArm64Main by getting
-        val macosArm64Test by getting
-        val linuxX64Main by getting
-        val linuxX64Test by getting
-        val linuxArm64Main by getting
-        val linuxArm64Test by getting
-        val iosMain by getting
-        val iosTest by getting
-        val iosSimulatorArm64Main by getting
-        val iosSimulatorArm64Test by getting
-        val watchosMain by getting
-        val watchosTest by getting
-        val watchosSimulatorArm64Main by getting
-        val watchosSimulatorArm64Test by getting
-        val tvosMain by getting
-        val tvosTest by getting
-        val tvosSimulatorArm64Main by getting
-        val tvosSimulatorArm64Test by getting
-
-        val nativeMain by sourceSets.creating {
-            dependsOn(commonMain)
-            linuxX64Main.dependsOn(this)
-            linuxArm64Main.dependsOn(this)
-        }
-        val nativeTest by sourceSets.creating {
-            dependsOn(commonTest)
-            linuxX64Test.dependsOn(this)
-            linuxArm64Test.dependsOn(this)
-        }
-
-        val appleMain by sourceSets.creating {
-            dependsOn(commonMain)
-            macosX64Main.dependsOn(this)
-            macosArm64Main.dependsOn(this)
-            iosMain.dependsOn(this)
-            iosSimulatorArm64Main.dependsOn(this)
-            watchosMain.dependsOn(this)
-            watchosSimulatorArm64Main.dependsOn(this)
-            tvosMain.dependsOn(this)
-            tvosSimulatorArm64Main.dependsOn(this)
-        }
-
-        val appleTest by sourceSets.creating {
-            dependsOn(commonTest)
-            macosX64Test.dependsOn(this)
-            macosArm64Test.dependsOn(this)
-            iosTest.dependsOn(this)
-            iosSimulatorArm64Test.dependsOn(this)
-            watchosTest.dependsOn(this)
-            watchosSimulatorArm64Test.dependsOn(this)
-            tvosTest.dependsOn(this)
-            tvosSimulatorArm64Test.dependsOn(this)
-        }
-
-        all {
-            languageSettings.optIn("kotlinx.cinterop.ExperimentalForeignApi")
-            languageSettings.optIn("kotlinx.cinterop.BetaInteropApi")
-            languageSettings.optIn("kotlinx.cinterop.UnsafeNumber")
+        jsMain.dependencies {
+            implementation("org.jetbrains.kotlin-wrappers:kotlin-web:1.0.0-pre.746")
+            implementation("org.jetbrains.kotlin-wrappers:kotlin-js:1.0.0-pre.746")
         }
     }
 }
