+++ src/androidInstrumentedTest/AndroidManifest.xml
