+++ src/androidMain/kotlin/com/ditchoom/buffer/Parcelable.kt
@@ -2,4 +2,5 @@ package com.ditchoom.buffer
 
 import android.os.Parcelable
 
+@Suppress("EXPECT_ACTUAL_CLASSIFIERS_ARE_IN_BETA_WARNING")
 actual typealias Parcelable = Parcelable
