+++ src/commonTest/kotlin/com/ditchoom/buffer/RunTest.kt
@@ -1,7 +1,3 @@
 package com.ditchoom.buffer
 
 expect fun <T> runTest(block: suspend () -> T)
-
-fun <T> runTestBlocking(block: suspend () -> T) {
-    runTest { block() }
-}
