+++ src/jvmMain/kotlin/com/ditchoom/buffer/Parcelable.kt
@@ -1,3 +1,4 @@
 package com.ditchoom.buffer
 
+@Suppress("EXPECT_ACTUAL_CLASSIFIERS_ARE_IN_BETA_WARNING")
 actual interface Parcelable
