+++ src/jvmTest/kotlin/com/ditchoom/buffer/RunTest.kt
@@ -1,5 +1,3 @@
-@file:JvmName("RunTest")
-
 package com.ditchoom.buffer
 
 import kotlinx.coroutines.runBlocking
