+++ src/nativeMain/kotlin/buffer/NativeBuffer.kt
