--- src/nativeMain/kotlin/com/ditchoom/buffer/Parcelable.kt
@@ -1,3 +1,0 @@
-package com.ditchoom.buffer
-
-actual interface Parcelable
